Added tabs container (HomePage) holding Rolodex Page, Exchange Page, and Profile Page w/ navigation; Added ion-card components as placeholders to serve as visuals for further development; *Further routing paths will need to be updated with the expanded layout
