+++ src/app/app-routing.module.ts
@@ -8,25 +8,18 @@ const routes: Routes = [
       pathMatch: 'full'
   },
   {
-    path: 'home',
-    loadChildren: './pages/home/home.module#HomePageModule'
+    path:'',
+    loadChildren:'./pages/home/home.module#HomePageModule'
   },
-  {
-    path: 'profile',
-    loadChildren: './pages/profile/profile.module#ProfilePageModule'
-  },
-  {
-    path: 'edit',
-    loadChildren: './pages/edit/edit.module#EditPageModule'
-  },
-  {
-    path: 'rolodex',
-    loadChildren: './pages/rolodex/rolodex.module#RolodexPageModule'
-  },
-  {
-    path: 'contact',
-    loadChildren: './pages/contact/contact.module#ContactPageModule'
-  },
+  //{
+    //path: 'edit',
+    //loadChildren: './pages/edit/edit.module#EditPageModule'
+  //},
+  //{
+    //path: 'contact',
+    //loadChildren: './pages/contact/contact.module#ContactPageModule'
+  //},
+
   { path: 'link', loadChildren: './pages/link/link.module#LinkPageModule' },
   { path: 'faq', loadChildren: './pages/faq/faq.module#FaqPageModule' },
   { path: 'about', loadChildren: './pages/about/about.module#AboutPageModule' },
@@ -34,7 +27,8 @@ const routes: Routes = [
   { path: 'settings', loadChildren: './pages/settings/settings.module#SettingsPageModule' },
   { path: 'feedback', loadChildren: './pages/feedback/feedback.module#FeedbackPageModule' },
   { path: 'invite', loadChildren: './pages/invite/invite.module#InvitePageModule' },
-  { path: 'password', loadChildren: './pages/password/password.module#PasswordPageModule' }
+  { path: 'password', loadChildren: './pages/password/password.module#PasswordPageModule' },
+  { path: 'exchange', loadChildren: './pages/home/exchange/exchange.module#ExchangePageModule' }
 
 ];
