+++ src/app/app.component.html
@@ -3,7 +3,7 @@
 </div>
 <ion-app>
   <ion-split-pane>
-    <ion-menu>
+    <ion-menu side="end">
       <ion-header>
         <ion-toolbar>
           <ion-title>Menu</ion-title>
