+++ src/app/app.component.ts
@@ -17,21 +17,21 @@ export class AppComponent {
   showSplash = true;
 
   public appPages = [
-    {
+      {
       title: 'Home',
       url: '/home',
       icon: 'home'
-    },
-    {
-      title: 'Profile',
-      url: '/profile',
-      icon: 'person'
-    },
-    {
-      title: 'Rolodex',
-      url: '/rolodex',
-      icon: 'aperture'
-    },
+      },
+    //{
+      //title: 'Profile',
+      //url: '/home/profile',
+      //icon: 'person'
+    //},
+    //{
+      //title: 'Rolodex',
+      //url: '/home/rolodex',
+      //icon: 'aperture'
+    //},
     {
       title: 'About',
       url: '/about',
