+++ src/app/pages/home/home.module.ts
@@ -2,21 +2,24 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
-import { RouterModule } from '@angular/router';
+import { Routes, RouterModule } from '@angular/router';
 
+import { HomePageRoutingModule } from './home-routing.module';
 import { HomePage } from './home.page';
+
+const routes: Routes = [
+  {
+    path:'',
+    component:HomePage
+  }
+];
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     IonicModule,
-    RouterModule.forChild([
-      {
-        path: '',
-        component: HomePage
-      }
-    ])
+    HomePageRoutingModule
   ],
   declarations: [HomePage]
 })
