+++ src/app/pages/home/profile/profile.page.html
@@ -1,0 +1,30 @@
+  <!--<ion-toolbar>
+    <ion-title>Profile</ion-title>
+  </ion-toolbar>-->
+
+  <ion-content>
+    <ion-card>
+    <ion-avatar class="image">
+      <img src="./assets/imgs/Beach.jpg" />
+    </ion-avatar>
+    <ion-card-header>
+      <ion-card-subtitle>Creative Artist</ion-card-subtitle>
+      <ion-card-title>Brandon Stuart</ion-card-title>
+   </ion-card-header>
+    </ion-card>
+  </ion-content>
+
+<ion-content>
+  <ion-list>
+    <ion-item *ngFor="let key of keys" lines="none">
+      {{key.toUpperCase()}} : {{user[key]}}
+    </ion-item>
+  </ion-list>
+
+
+  <ion-fab vertical="bottom" horizontal="start" slot="fixed">
+    <ion-fab-button (click)="goToEditPage()">
+      <ion-icon name="add"></ion-icon>
+    </ion-fab-button>
+  </ion-fab>
+</ion-content>
