+++ src/app/pages/home/profile/profile.page.scss
@@ -1,0 +1,6 @@
+.image {
+  height: 75px;
+  width: 75px;
+  padding-top: 5px;
+  padding-left: 5px;
+}
