+++ src/app/pages/home/profile/profile.page.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+selector: 'app-profile',
+templateUrl: './profile.page.html',
+styleUrls: ['./profile.page.scss'],
+})
+export class ProfilePage implements OnInit {
+
+constructor() { }
+
+  ngOnInit() {
+  }
+
+}
