+++ src/app/pages/home/rolodex/rolodex.page.html
@@ -1,0 +1,73 @@
+ <ion-toolbar>
+    <ion-buttons slot="end">
+      <ion-button (click)="addContact()">
+        <ion-icon name="person-add"></ion-icon>
+      </ion-button>
+    </ion-buttons>
+    <!--<ion-title>
+      Rolodex
+    </ion-title>-->
+  </ion-toolbar>
+
+<ion-content>
+
+  <ion-card>
+
+    <ion-card-header>
+      <ion-card-subtitle>Card Subtitle</ion-card-subtitle>
+      <ion-card-title>Card Title</ion-card-title>
+    </ion-card-header>
+
+    <ion-card-content>
+    </ion-card-content>
+  </ion-card>
+
+  <ion-card>
+    <ion-card-header>
+      <ion-card-subtitle>Card Subtitle</ion-card-subtitle>
+      <ion-card-title>Card Title</ion-card-title>
+    </ion-card-header>
+
+    <ion-card-content>
+    </ion-card-content>
+
+  </ion-card>
+
+  <ion-card>
+  <ion-card-header>
+    <ion-card-subtitle>Card Subtitle</ion-card-subtitle>
+    <ion-card-title>Card Title</ion-card-title>
+  </ion-card-header>
+
+  <ion-card-content>
+  </ion-card-content>
+
+</ion-card>
+
+  <ion-card>
+    <ion-card-header>
+      <ion-card-subtitle>Card Subtitle</ion-card-subtitle>
+      <ion-card-title>Card Title</ion-card-title>
+    </ion-card-header>
+
+    <ion-card-content>
+    </ion-card-content>
+
+  </ion-card>
+
+
+
+  <ion-list>
+    <ion-item *ngFor="let contactData of contactDataList" lines="none">
+      {{contactData.name}}
+      <ion-buttons slot="end">
+        <ion-button (click)="goToContactPage(contactData.id)">
+          <ion-icon name="eye"></ion-icon>
+        </ion-button>
+        <ion-button (click)="deleteContact(contactData.id)">
+          <ion-icon name="trash"></ion-icon>
+        </ion-button>
+      </ion-buttons>
+    </ion-item>
+  </ion-list>
+</ion-content>
