+++ src/app/pages/home/rolodex/rolodex.page.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+selector: 'app-rolodex',
+templateUrl: './rolodex.page.html',
+styleUrls: ['./rolodex.page.scss'],
+})
+export class RolodexPage implements OnInit {
+
+constructor() { }
+
+  ngOnInit() {
+  }
+
+}
