--- src/app/pages/profile/profile.page.html
@@ -1,21 +1,0 @@
-<ion-header>
-  <ion-toolbar>
-    <ion-buttons slot="start">
-      <ion-menu-button></ion-menu-button>
-    </ion-buttons>
-    <ion-title>Profile</ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content>
-  <ion-list>
-    <ion-item *ngFor="let key of keys" lines="none">
-      {{key.toUpperCase()}} : {{user[key]}}
-    </ion-item>
-  </ion-list>
-
-  <ion-button (click)="goToEditPage()">
-    Edit
-  </ion-button>
-
-</ion-content>
