--- src/app/pages/rolodex/rolodex.page.html
@@ -1,33 +1,0 @@
-<ion-header>
-  <ion-toolbar>
-    <ion-buttons slot="start">
-      <ion-menu-button></ion-menu-button>
-
-    </ion-buttons>
-    <ion-buttons slot="end">
-      <ion-button (click)="addContact()">
-        <ion-icon name="person-add"></ion-icon>
-      </ion-button>
-    </ion-buttons>
-    <ion-title>
-      Rolodex
-    </ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content>
-
-  <ion-list>
-    <ion-item *ngFor="let contactData of contactDataList" lines="none">
-      {{contactData.name}}
-      <ion-buttons slot="end">
-        <ion-button (click)="goToContactPage(contactData.id)">
-          <ion-icon name="eye"></ion-icon>
-        </ion-button>
-        <ion-button (click)="deleteContact(contactData.id)">
-          <ion-icon name="trash"></ion-icon>
-        </ion-button>
-      </ion-buttons>
-    </ion-item>
-  </ion-list>
-</ion-content>
